Use conan to download catch2 on CMake configure
Write a conanfile.txt file in the root directory of the project, select
cmake as the generator. Run the conan install command on it at CMake
configure time. This will download catch2 (or any libary in the
conanfile) and output a .cmake file. We can include this .cmake file
which will contain the paths for the libraries, so we can link to them.